refactor: extract SessionRolloverDetector from LcmContextEngine
Phase 3 of the engine.ts decomposition (#867). The missed-lifecycle
rollover cluster (stale session-key transcript loss, isolated-cron
runtime changes, ambiguous session-key rollover freshness evaluation and
rotation) moves to src/session-rollover.ts. The detector takes the two
stores, the logger, and an applySessionReplacement callback (session
replacement itself stays engine-side with the session queue). Also drops
a SummarySpendGuardState type left behind by the Phase 2 guards move.

Full suite: 1279 tests passing.
